+++ app-gui/src/main/java/com/microej/demo/sandbox/ui/style/AppColors.java
@@ -13,8 +13,12 @@ public class AppColors {
 
 	/** Black color for the content background. */
 	public static final int CONTENT_BG_BLACK = 0x1d1d1d;
-	/** Orange color for the progress line fill. */
-	public static final int PROGRESS_LINE_FOREGROUND = 0xf04f25;
+	/** Gradient min range color for the progress line fill (Green). */
+	public static final int PROGRESS_LINE_FOREGROUND_MIN = 0x6CC24A;
+	/** Gradient max range color for the progress line fill (Yellow). */
+	public static final int PROGRESS_LINE_FOREGROUND_MID = 0xFFC845;
+	/** Gradient max range color for the progress line fill (Orange). */
+	public static final int PROGRESS_LINE_FOREGROUND_MAX = 0xEE502E;
 	/** Gray color for the progress line bg. */
 	public static final int PROGRESS_LINE_BACKGROUND = 0x4e4e4e;
 	/** Gray color for the secondary label fill. */