+++ app-mqtt-publisher/src/main/java/com/microej/demo/sandbox/publisher/MqttPublisher.java
@@ -1,7 +1,7 @@
 /*
  * Java
  *
- * Copyright 2023-2024 MicroEJ Corp. All rights reserved.
+ * Copyright 2023-2025 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.demo.sandbox.publisher;
@@ -10,11 +10,12 @@ import java.util.Random;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import com.microej.demo.sandbox.sharedinterface.PowerServiceListenerUtil;
 import org.eclipse.paho.client.mqttv3.MqttClient;
 import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 
-import com.microej.demo.sandbox.sharedinterface.Observer;
+import com.microej.demo.sandbox.sharedinterface.PowerServiceListener;
 import com.microej.demo.sandbox.sharedinterface.PowerService;
 
 import ej.service.ServiceFactory;
@@ -22,10 +23,10 @@ import ej.service.ServiceFactory;
 /**
  * The MqttPublisher class publishes the power readings from the PowerProvider into an MQTT queue.
  */
-public class MqttPublisher implements Observer {
+public class MqttPublisher implements PowerServiceListener {
 
 	private static final Logger LOGGER = Logger.getLogger("MqttPublisher"); //$NON-NLS-1$
-	private static final Random RANDOM = new Random();
+	private static final Random RANDOM = new Random(); // NOSONAR the random is generated for the demo purpose, it is not unsafe to use it
 
 	/**
 	 * URL of the MQTT broker.
@@ -42,6 +43,7 @@ public class MqttPublisher implements Observer {
 	 */
 	public static final String TOPIC_POWER = "microej/demo/sandbox/power_" + RANDOM.nextInt(Integer.MAX_VALUE); //$NON-NLS-1$
 
+	private final PowerServiceListenerUtil powerServiceListenerUtil = new PowerServiceListenerUtil();
 	private final MqttClient client = new MqttClient(BROKER, PUBLISHER_ID);
 	private final PowerSubscriber powerSubscriber = new PowerSubscriber();
 	private int lastValue = -1;
@@ -56,6 +58,8 @@ public class MqttPublisher implements Observer {
 
 			this.client.subscribe(TOPIC_POWER);
 			this.client.setCallback(this.powerSubscriber);
+			PowerService powerService = this.powerServiceListenerUtil.waitAndGetPowerService();
+			powerService.addListener(this);
 		} catch (MqttException e) {
 			LOGGER.log(Level.SEVERE, "MQTT ERROR", e); //$NON-NLS-1$
 		}
@@ -76,15 +80,20 @@ public class MqttPublisher implements Observer {
 
 	@Override
 	public void update() {
-		if (!MqttPublisher.this.client.isConnected()) {
-			LOGGER.severe("Client is not connected."); //$NON-NLS-1$
-			return;
-		}
-
 		// Get current data from data provider
-		PowerService powerService = ServiceFactory.getService(PowerService.class);
+		final PowerService powerService = ServiceFactory.getService(PowerService.class);
 		if (powerService == null) {
-			LOGGER.severe("PowerManager service not found."); //$NON-NLS-1$
+			LOGGER.warning("MeterData service not found."); //$NON-NLS-1$
+
+			// PowerService was not found, wait for it and listen to it again
+			new Thread(new Runnable() {				// NOSONAR due to a bug in the sonar plugin, the java sources are
+													// set to java 8 and this anonymous class is considered as a code smell
+				@Override
+				public void run() {
+					MqttPublisher.this.powerServiceListenerUtil.waitAndGetPowerService().addListener(MqttPublisher.this);
+				}
+			}).start();
+
 			return;
 		}
 
@@ -96,7 +105,7 @@ public class MqttPublisher implements Observer {
 	}
 
 	/**
-	 * Sends an message with the provided power.
+	 * Sends a message with the provided power.
 	 *
 	 * @param message
 	 *            the message
@@ -105,8 +114,8 @@ public class MqttPublisher implements Observer {
 		MqttMessage mqttMessage = new MqttMessage(message.getBytes());
 		try {
 			if (!this.client.isConnected()) {
-				LOGGER.log(Level.WARNING, "Client is not connected"); //$NON-NLS-1$
-				return;
+				LOGGER.log(Level.WARNING, "Client is not connected, trying to reconnect"); //$NON-NLS-1$
+				this.connect();
 			}
 			this.client.publish(MqttPublisher.TOPIC_POWER, mqttMessage);
 		} catch (MqttException e) {