+++ app-power-provider/src/main/java/com/microej/demo/sandbox/provider/PowerProviderEntryPoint.java
@@ -1,7 +1,7 @@
 /*
  * Java
  *
- * Copyright 2023-2024 MicroEJ Corp. All rights reserved.
+ * Copyright 2023-2025 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.demo.sandbox.provider;