+++ CHANGELOG.rst
@@ -7,6 +7,17 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+------------------
+2.0.0 - 2025-04-10
+------------------
+
+- Migrate project to SDK6.
+- Update default kernel to GREEN kernel 2.1.1 on MIMXRT1170 v3.0.0 VEE Port.
+- Update shared interface example project.
+- Update provider example project.
+- Update UI example project.
+- Update publisher example project.
+
 -------------------
  1.0.0 - 2024-05-30
 -------------------
@@ -21,5 +32,5 @@ Added
 - Add publisher example project.
 
 .. ReStructuredText
-.. Copyright 2023-2024 MicroEJ Corp. All rights reserved.
+.. Copyright 2023-2025 MicroEJ Corp. All rights reserved.
 .. Use of this source code is governed by a BSD-style license that can be found with this software.