+++ app-power-provider/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2025-04-10
+
+### Changed
+
+- Migrate project to SDK6.
+- Remove deprecated observer class usage.
+
 ## [1.0.0] - 2024-05-30
 
 ### Added
@@ -14,5 +21,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ---  
-_Copyright 2023-2024 MicroEJ Corp. All rights reserved._  
+_Copyright 2023-2025 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._